requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
alembic
